backend/requirements.txt
django==5.2.12
djangorestframework==3.16.1
requests==2.32.5
psycopg2-binary==2.9.11
python-dotenv==1.2.2
